Snap: Use HAL3 mode for the OnePlus 3T's front-facing camera
There is a bug in OnePlus' camera HAL that causes the camera to crash when
capturing pictures with the front-facing camera that are larger than 8.0 MP.
This bug only exists in HAL1 mode; HAL3 mode is immune.

The OnePlus 3's front-facing camera's maximum resolution is conveniently
8.0 MP, so the OnePlus 3 is unaffected. Therefore, force HAL3 mode for the
front-facing camera of only the OnePlus 3T.

The camcorder is fine though, so make the camcorder continue to use HAL1 mode.