Snap: Fix low exposure bug when using OP3T's FFC to record video
There is a bug in the camera HAL that affects the OnePlus 3T's front-facing
camera in camcorder mode. The bug causes the exposure to be unusably low,
making videos recorded with the front-facing camera appear nearly black.

However, the bug does not appear when the rear camera is opened in
camcorder mode first. Therefore, to work around this bug, change the
user flow to make camcorder mode always open the rear camera when
transitioning from photo mode. That way, the rear camera will _always_
be opened in camcorder mode before the front camera can.