Snap: Attempt to fix camera actuator not working
Sometimes when the camera is opened through Snap, the actuator refuses to
work, causing the camera to be stuck at a fixed focal position (just like
the "fixed" focus mode). The only way to make the camera start focusing
correctly again in Snap is to force close the app and re-open it.

In an attempt to make the actuator work always work when it's supposed to,
do something similar to the ISO hack by changing the focus mode to Fixed
and then back to CAF.