Snap: Always perform autofocus before capture using focus areas
The OP3(T)'s continuous autofocus configuration is very buggy and
constantly throws the camera out of focus. This causes many pictures taken
to come out blurry and out of focus. This bug is in the BSP, so we cannot
truly fix it.

To improve the focus situation, force an autofocus event every time a
photo is captured while continuous autofocus is in use.

Note that simply changing the focus mode to auto before capture and relying
on the camera HAL to perform the autofocus event for us is not safe. This
causes the camera to frequently freeze when attempting to capture photos
with flash, so we have to use focus areas to achieve this instead.