Snap: Hide preview during rear -> front cam transition
When the front-facing camera is opened in HAL3 mode, there is a bug in
the camera HAL that causes the camera preview to appear distorted for
about half a second when switching from the rear camera to the front-facing
camera. In order to prevent the user from seeing the ugly graphical
glitch, hide the camera preview for about 400 ms during the rear-to-front
camera transition. 400 ms is roughly 100 ms longer than the time it takes
for the graphical glitch to disappear.